requirements.txt
flask
tensorflow
pillow
umpy
